flask-app/requirements.txt
flask
Flask-SocketIO
flask_sqlalchemy
flask-login
stripe
boto3
pillow
python-dotenv